dedup get_affected_registers
