+++ crates/mipsy_interactive/src/interactive/mod.rs
@@ -6,7 +6,7 @@ mod runtime_handler;
 
 use std::{mem::take, ops::Deref, rc::Rc, sync::{atomic::{AtomicBool, Ordering}, Arc}};
 
-use mipsy_lib::{Binary, InstSet, Runtime, MipsyError, ParserError, error::parser, runtime::{SteppedRuntime, state::TIMELINE_MAX_LEN, SPECIAL, SPECIAL2, SPECIAL3, JUMP, JAL}, Register, compile::breakpoints::{WatchpointTarget, TargetWatch, TargetAction}};
+use mipsy_lib::{Binary, InstSet, Runtime, MipsyError, ParserError, error::parser, runtime::{SteppedRuntime, state::TIMELINE_MAX_LEN}, compile::breakpoints::{TargetWatch, TargetAction, get_affected_registers}};
 use mipsy_lib::runtime::{SYS13_OPEN, SYS14_READ, SYS15_WRITE, SYS16_CLOSE};
 use mipsy_lib::error::runtime::{Error, RuntimeError, ErrorContext, InvalidSyscallReason};
 use helper::MyHelper;
@@ -31,17 +31,6 @@ use commands::{
 use mipsy_utils::MipsyConfig;
 
 use self::error::{CommandError, CommandResult};
-
-const LB : u32 = 0b100000;
-const LBU: u32 = 0b100100;
-const LH : u32 = 0b100001;
-const LHU: u32 = 0b100101;
-const LUI: u32 = 0b001111;
-const LW : u32 = 0b100011;
-const LWU: u32 = 0b100111;
-const SB : u32 = 0b101000;
-const SH : u32 = 0b101001;
-const SW : u32 = 0b101011;
 
 pub(crate) struct State {
     pub(crate) config: MipsyConfig,
@@ -464,7 +453,7 @@ impl State {
 
         let mut empty_binary = Binary::default();
         let binary = self.binary.as_mut().unwrap_or(&mut empty_binary);
-        let affected_registers = Self::get_affected_registers(&self.runtime, inst);
+        let affected_registers = get_affected_registers(&self.runtime, inst);
         // TODO(joshh): move this into else if once let-chains are stabilised (1.64 baited me smh)
         let watchpoints = affected_registers.iter()
             .filter(|&wp| binary.watchpoints.get(&wp.target)
@@ -523,79 +512,6 @@ impl State {
                 }
             }
         )
-    }
-
-    fn get_affected_registers(runtime: &Runtime, inst: u32) -> Vec<TargetWatch> {
-        let opcode =  inst >> 26;
-        let rb     = (inst >> 21) & 0x1F;
-        let rs     = (inst >> 21) & 0x1F;
-        let rt     = (inst >> 16) & 0x1F;
-        let rd     = (inst >> 11) & 0x1F;
-        let offset = (inst & 0xFF) as i32;
-
-        match opcode {
-            LB | LBU | LH | LHU | LW | LWU | LUI => vec![
-                TargetWatch {
-                    target: WatchpointTarget::Register(Register::from_u32(rt).unwrap()),
-                    action: TargetAction::WriteOnly,
-                },
-                TargetWatch {
-                    target: WatchpointTarget::Register(Register::from_u32(rb).unwrap()),
-                    action: TargetAction::ReadOnly,
-                },
-                TargetWatch {
-                    target: WatchpointTarget::MemAddr(
-                        (runtime.timeline().prev_state().expect("there should be a previous state")
-                            .read_register(rb).expect("uninitialized read should already have been handled")
-                            + offset) as u32
-                    ),
-                    action: TargetAction::ReadOnly,
-                },
-            ],
-            SB | SH | SW => vec![
-                TargetWatch {
-                    target: WatchpointTarget::MemAddr(
-                        (runtime.timeline().state()
-                            .read_register(rb).expect("uninitialized read should already have been handled")
-                            + offset) as u32
-                    ),
-                    action: TargetAction::WriteOnly,
-                },
-                TargetWatch {
-                    target: WatchpointTarget::Register(Register::from_u32(rt).unwrap()),
-                    action: TargetAction::ReadOnly,
-                },
-                TargetWatch {
-                    target: WatchpointTarget::Register(Register::from_u32(rb).unwrap()),
-                    action: TargetAction::ReadOnly,
-                },
-            ],
-            SPECIAL | SPECIAL2 | SPECIAL3 => vec![
-                TargetWatch {
-                    target: WatchpointTarget::Register(Register::from_u32(rd).unwrap()),
-                    action: TargetAction::WriteOnly,
-                },
-                TargetWatch {
-                    target: WatchpointTarget::Register(Register::from_u32(rs).unwrap()),
-                    action: TargetAction::ReadOnly,
-                },
-                TargetWatch {
-                    target: WatchpointTarget::Register(Register::from_u32(rt).unwrap()),
-                    action: TargetAction::ReadOnly,
-                },
-            ],
-            JUMP | JAL => vec![],
-            _ => vec![
-                TargetWatch {
-                    target: WatchpointTarget::Register(Register::from_u32(rs).unwrap()),
-                    action: TargetAction::ReadOnly,
-                },
-                TargetWatch {
-                    target: WatchpointTarget::Register(Register::from_u32(rt).unwrap()),
-                    action: TargetAction::WriteOnly,
-                },
-            ],
-        }
     }
 
     pub(crate) fn step(&mut self, verbose: bool) -> CommandResult<bool> {
