+++ crates/mipsy_lib/src/compile/breakpoints.rs
@@ -1,7 +1,7 @@
 use colored::Colorize;
 use serde::{Deserialize, Serialize};
 
-use crate::Register;
+use crate::{Register, Runtime, runtime::{SPECIAL, SPECIAL2, SPECIAL3, JAL, JUMP}};
 use std::fmt::Display;
 
 #[derive(Copy, Clone, Debug, Serialize, Deserialize)]
@@ -123,5 +123,90 @@ impl Point for Watchpoint {
 
     fn get_commands(&'_ mut self) -> &'_ mut Vec<String> {
         &mut self.commands
+    }
+}
+
+
+const LB : u32 = 0b100000;
+const LBU: u32 = 0b100100;
+const LH : u32 = 0b100001;
+const LHU: u32 = 0b100101;
+const LUI: u32 = 0b001111;
+const LW : u32 = 0b100011;
+const LWU: u32 = 0b100111;
+const SB : u32 = 0b101000;
+const SH : u32 = 0b101001;
+const SW : u32 = 0b101011;
+
+pub fn get_affected_registers(runtime: &Runtime, inst: u32) -> Vec<TargetWatch> {
+    let opcode =  inst >> 26;
+    let rb     = (inst >> 21) & 0x1F;
+    let rs     = (inst >> 21) & 0x1F;
+    let rt     = (inst >> 16) & 0x1F;
+    let rd     = (inst >> 11) & 0x1F;
+    let offset = (inst & 0xFF) as i32;
+
+    match opcode {
+        LB | LBU | LH | LHU | LW | LWU | LUI => vec![
+            TargetWatch {
+                target: WatchpointTarget::Register(Register::from_u32(rt).unwrap()),
+                action: TargetAction::WriteOnly,
+            },
+            TargetWatch {
+                target: WatchpointTarget::Register(Register::from_u32(rb).unwrap()),
+                action: TargetAction::ReadOnly,
+            },
+            TargetWatch {
+                target: WatchpointTarget::MemAddr(
+                    (runtime.timeline().prev_state().expect("there should be a previous state")
+                        .read_register(rb).expect("uninitialized read should already have been handled")
+                        + offset) as u32
+                ),
+                action: TargetAction::ReadOnly,
+            },
+        ],
+        SB | SH | SW => vec![
+            TargetWatch {
+                target: WatchpointTarget::MemAddr(
+                    (runtime.timeline().state()
+                        .read_register(rb).expect("uninitialized read should already have been handled")
+                        + offset) as u32
+                ),
+                action: TargetAction::WriteOnly,
+            },
+            TargetWatch {
+                target: WatchpointTarget::Register(Register::from_u32(rt).unwrap()),
+                action: TargetAction::ReadOnly,
+            },
+            TargetWatch {
+                target: WatchpointTarget::Register(Register::from_u32(rb).unwrap()),
+                action: TargetAction::ReadOnly,
+            },
+        ],
+        SPECIAL | SPECIAL2 | SPECIAL3 => vec![
+            TargetWatch {
+                target: WatchpointTarget::Register(Register::from_u32(rd).unwrap()),
+                action: TargetAction::WriteOnly,
+            },
+            TargetWatch {
+                target: WatchpointTarget::Register(Register::from_u32(rs).unwrap()),
+                action: TargetAction::ReadOnly,
+            },
+            TargetWatch {
+                target: WatchpointTarget::Register(Register::from_u32(rt).unwrap()),
+                action: TargetAction::ReadOnly,
+            },
+        ],
+        JUMP | JAL => vec![],
+        _ => vec![
+            TargetWatch {
+                target: WatchpointTarget::Register(Register::from_u32(rs).unwrap()),
+                action: TargetAction::ReadOnly,
+            },
+            TargetWatch {
+                target: WatchpointTarget::Register(Register::from_u32(rt).unwrap()),
+                action: TargetAction::WriteOnly,
+            },
+        ],
     }
 }
